scripts/requirements.txt
# For scripts/clear_postgres_chroma.py
psycopg2-binary>=2.9.9
chromadb>=0.5.0
